+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/actions/share/impl/ShareServlet.java
@@ -23,11 +23,11 @@ import com.adobe.aem.commons.assetshare.components.actions.share.ShareException;
 import com.adobe.aem.commons.assetshare.components.actions.share.ShareService;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.SlingHttpServletResponse;
-import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.api.servlets.SlingAllMethodsServlet;
 import org.apache.sling.api.wrappers.ValueMapDecorator;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferencePolicyOption;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,20 +35,28 @@ import javax.jcr.RepositoryException;
 import javax.servlet.Servlet;
 import javax.servlet.ServletException;
 import java.io.IOException;
+import java.util.Collection;
 import java.util.HashMap;
+import java.util.Objects;
+import java.util.concurrent.atomic.AtomicInteger;
 
-@Component(service = Servlet.class, property = {
-        "sling.servlet.methods=POST",
-        "sling.servlet.resourceTypes=asset-share-commons/actions/share",
-        "sling.servlet.selectors=share",
-        "sling.servlet.extensions=html",
-        "sling.servlet.extensions=json"
-})
+@Component(service = Servlet.class,
+           property = {
+                   "sling.servlet.methods=POST",
+                   "sling.servlet.resourceTypes=asset-share-commons/actions/share",
+                   "sling.servlet.selectors=share",
+                   "sling.servlet.extensions=html",
+                   "sling.servlet.extensions=json"
+           }
+)
 public class ShareServlet extends SlingAllMethodsServlet {
     private static final Logger log = LoggerFactory.getLogger(ShareServlet.class);
 
-    @Reference
-    private ShareService shareService;
+    @Reference(target = "(component.name=com.adobe.aem.commons.assetshare.components.actions.share.impl.EmailShareServiceImpl)")
+    private ShareService defaultShareService;
+
+    @Reference(policyOption = ReferencePolicyOption.GREEDY)
+    private transient Collection<ShareService> shareServices;
 
     @Override
     protected final void doPost(SlingHttpServletRequest request, SlingHttpServletResponse response)
@@ -61,14 +69,31 @@ public class ShareServlet extends SlingAllMethodsServlet {
     }
 
     private final void share(SlingHttpServletRequest request, SlingHttpServletResponse response) throws RepositoryException {
+
+        // Make this map write-able by copying the Request parameters into a new hash map
+        final AtomicInteger counter = new AtomicInteger(0);
+
+        // Call all accepting ShareService implementations
+        shareServices.stream()
+                .filter(Objects::nonNull)
+                .filter(shareService -> shareService.accepts(request))
+                .forEach(shareService -> {
+                    try {
+                        shareService.share(request, response,
+                                // Make map write-able
+                                new ValueMapDecorator(new HashMap<String, Object>(request.getParameterMap())));
+                        counter.incrementAndGet();
+                    } catch (ShareException e) {
+                        log.error("Failed to not share assets using [ {} ]", shareService.getClass().getName(), e);
+                    }
+                });
+
         try {
-            if (shareService != null) {
-                // Make this map write-able by copying the Request parameters into a new hash map
-                final ValueMap shareParameters = new ValueMapDecorator(new HashMap<String, Object>(request.getParameterMap()));
-                // Call the best ranking Share Service
-                shareService.share(request, response, shareParameters);
-            } else {
-                throw new ShareException("No ShareService is active, so sharing is disabled");
+            // If no accepting implementations can be found, use the default ASC Email implementation.
+            if (counter.get() == 0) {
+                defaultShareService.share(request, response,
+                        // Make map write-able
+                        new ValueMapDecorator(new HashMap<String, Object>(request.getParameterMap())));
             }
         } catch (ShareException ex) {
             log.error("Unable to share assets from Asset Share Commons", ex);