+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/actions/share/package-info.java
@@ -17,7 +17,7 @@
  *
  */
 
-@Version("1.0.1")
+@Version("2.0.0")
 package com.adobe.aem.commons.assetshare.components.actions.share;
 
 import org.osgi.annotation.versioning.Version;