+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - 0128: Path Filter search component.
 - 0130: Added auto-search capabilities to search predicate components.
+- 0134: Allow multiple ShareServices to be registered and allow each to accept the request.
 
 ## [v1.2.2]